Decouple metrics clients and validators into their own modules
